requirements.txt
numpy
tensorflow==2.16.1
Pillow
streamlit